updated tests

+++ src/main/java/RewardValue.java
@@ -3,14 +3,14 @@ import java.util.Scanner;
 public class RewardValue {
 
     private double cashValue;    // cash value
-    private double milesValue;     // amount of miles
+    private int milesValue;     // amount of miles
 
     public RewardValue(double args) {
         // RewardValue converts from miles to cash at a rate of 0.0035.
         // set the amount of cash that the user has entered
 
         this.cashValue = args;
-        this.milesValue = args * 0.0035;
+        this.milesValue = (int)(args * 0.0035);
     }
 
     public RewardValue(int args) {
@@ -18,7 +18,7 @@ public class RewardValue {
         // set the amount of cash that the user has entered
 
         this.milesValue = args;
-        this.cashValue = args / 0.0035;
+        this.cashValue = (double)(args / 0.0035);
     }
 
     public double getCashValue() {
